+++ src/libs/Notification.ts
@@ -40,6 +40,7 @@ export default class Notification implements Serializable<SerializedNotification
  * Notification associated events.
  */
 export enum NotificationEvent {
+  Host,
   Raid,
   ReSub,
   RitualNewChatter,